+++ test/snippet/utility/views/type_reduce.cpp
@@ -2,7 +2,7 @@
 #include <vector>
 
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/type_reduce.hpp>
+#include <seqan3/utility/views/type_reduce.hpp>
 
 int main()
 {